backend/rag_api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
openai>=1.12.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
neo4j>=5.14.0

